Code/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
python-socketio==5.10.0
python-engineio==4.8.0
WTForms==3.1.1
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
pymssql==2.3.1
SQLAlchemy>=2.0.36
